Zmiana ToIn32 na ToByte

+++ BMI/Program.cs
@@ -21,7 +21,7 @@ namespace BMI
             int waga = Convert.ToInt32(Console.ReadLine());
 
             Console.Write("wprowadz wzrost w cm:");
-            int wzrost = Convert.ToInt32(Console.ReadLine());
+            byte wzrost = Convert.ToByte(Console.ReadLine());
 
             double bmi = waga / Math.Pow(wzrost / 100.0, 2);
             Console.WriteLine("BMI="+bmi);
